Correção codesmell 1.

+++ src/main/java/com/dynss/cloudtecnologia/rest/dto/DashboardDTO.java
@@ -2,15 +2,13 @@ package com.dynss.cloudtecnologia.rest.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 public class DashboardDTO {
     List<LancamentoReflectionDTO> lancamentos;
-    BigDecimal sumEntradas;
-    BigDecimal sumSaidas;
+    long sumEntradas;
+    long sumSaidas;
     Integer ano;
 }
